requirements.txt
blessed==1.20.0
colorlog==6.8.2
editor==1.6.6
inquirer==3.4.0
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.2
readchar==4.2.0
rich==13.8.1
runs==1.2.2
six==1.16.0
wcwidth==0.2.13
xmod==1.8.1
